requirements.txt
opencv-contrib-python
numpy
scipy
h5py
matplotlib
pytorch
torchvision